Workers now have user-defined hours, task durations rebalanced, #2 #6 related

+++ Assets/Scripts/ActionItem.cs
@@ -20,7 +20,7 @@ public abstract class ActionItem : ScriptableObject
     public abstract float TimeRemaining();
     public abstract bool NeedWorkers(); // If true, the current step needs workers
     public abstract void AutoStep(); // Move to next step if current complete
-    public abstract int QtyTime(); // Amount of time it takes to process 1 quantity
+    public abstract float QtyTime(); // Amount of time it takes to process 1 quantity
     public abstract void ProcessStock(int qty); // Process the specified qty in current step
     public abstract bool CanDoStep(int stockRacked); // Enough stock to pick for an order
     public abstract string ForwardText(); // Text to display on "accept" button for current step
